Allow creation of Metric from datastructure without __name__
When applying a function to a time series in a query to Prometheus,
the Prometheus server returns a data structure without __name__ key.
Apart from that aspect, the time series could still be used to create
a Metric object.

This PR proposes to set "metric_name" to None when __name__ is missing.